Dto added

+++ lms/src/main/java/com/project/lms/Service/BooksService.java
@@ -1,12 +1,16 @@
 package com.project.lms.Service;
 
 
+import com.project.lms.Dto.ApiResponse;
+import com.project.lms.Dto.BooksRequest;
+import com.project.lms.Dto.BooksResponse;
 import com.project.lms.Entity.Books;
 import com.project.lms.Repository.BooksRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class BooksService {
@@ -14,25 +18,71 @@ public class BooksService {
     @Autowired
     private BooksRepository booksRepository;
 
-    public List<Books> getAllBooks() {
-        return booksRepository.findAll();
+    private BooksResponse toBookResponse(Books books){
+        return new BooksResponse(
+                books.getIsbn(), books.getTitle(), books.getEdition(), books.getAuthName(), books.getPrice(), books.getCategory()
+        );
     }
 
-    public Books getBookByIsbn(String isbn) {
-        return booksRepository.findById(isbn).orElse(null);
+    public ApiResponse<List<BooksResponse>> getAllBooks() {
+        List<Books> books = booksRepository.findAll();
+        if(books.isEmpty())
+            return ApiResponse.fail("No books available currently.");
+        List<BooksResponse> res = books.stream().map(this::toBookResponse).toList();
+        return ApiResponse.ok("All Books Fetched.", res);
     }
 
-    public Books addBook(Books book) {
-        return booksRepository.save(book);
+    public ApiResponse<BooksResponse> getBookByIsbn(String isbn) {
+        Optional<Books> booksOptional = booksRepository.findById(isbn);
+        if(booksOptional.isEmpty())
+            return ApiResponse.fail("Book not found.");
+        Books books = booksOptional.get();
+        return ApiResponse.ok("Book Fetched.", toBookResponse(books));
     }
 
-    public Books updateBook(String isbn, Books book) {
-
-        Books existingBook = booksRepository.findById(isbn).orElse(null);
-        return null;
+    public ApiResponse<BooksResponse> addBook(BooksRequest bookreq) {
+        if(bookreq.getAuthName().isBlank() || bookreq.getTitle().isBlank() || bookreq.getIsbn().isBlank() || bookreq.getCategory().isBlank() || bookreq.getPrice() != 0)
+            return ApiResponse.fail("All fields are required.");
+        if(!bookreq.getIsbn().equals(bookreq.getConfirm_isbn()))
+            return ApiResponse.fail("Isbn does not match.");
+        Optional<Books> booksOptional = booksRepository.findById(bookreq.getIsbn());
+        if(booksOptional.isPresent())
+            return ApiResponse.fail("Book with this Isbn already exists.");
+        Books books = new Books();
+        books.setIsbn(bookreq.getIsbn());
+        books.setTitle(bookreq.getTitle());
+        books.setEdition(bookreq.getEdition());
+        books.setAuthName(bookreq.getAuthName());
+        books.setPrice(bookreq.getPrice());
+        books.setCategory(bookreq.getCategory());
+        booksRepository.save(books);
+        return ApiResponse.ok("Book Added.", toBookResponse(books));
     }
 
-    public void deleteBook(String isbn) {
+    public ApiResponse<BooksResponse> updateBook(String isbn, BooksRequest bookreq) {
+        Optional<Books> bookOptional = booksRepository.findById(isbn);
+        if(bookOptional.isEmpty())
+            return ApiResponse.fail("Book not found.");
+        Books existingBook = bookOptional.get();
+        if(!bookreq.getTitle().isBlank())
+            existingBook.setTitle(bookreq.getTitle());
+        if(!bookreq.getAuthName().isBlank())
+            existingBook.setAuthName(bookreq.getAuthName());
+        if(bookreq.getEdition() != 0)
+            existingBook.setEdition(bookreq.getEdition());
+        if(!bookreq.getCategory().isBlank())
+            existingBook.setCategory(bookreq.getCategory());
+        if(bookreq.getPrice() != 0)
+            existingBook.setPrice(bookreq.getPrice());
+        booksRepository.save(existingBook);
+        return ApiResponse.ok("Book updated.", toBookResponse(existingBook));
+    }
+
+    public ApiResponse<Void> deleteBook(String isbn) {
+        Optional<Books> booksOptional = booksRepository.findById(isbn);
+        if(booksOptional.isEmpty())
+            return ApiResponse.fail("Book not found.");
         booksRepository.deleteById(isbn);
+        return ApiResponse.ok("Book Deleted", null);
     }
 }
